feat(runtime): port upstream lazy layout storage and hit ordering
Upstream-Commit: 61f00897013eba8c57ee2002c33e6ee47d476eff

+++ crates/nuxie-runtime/src/mechanical_port/source/animation/state_machine_fire_action.rs
@@ -12,7 +12,7 @@ use crate::mechanical_port::source::{
 };
 
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub struct StateMachineFireOccurance(pub i32);
+pub struct StateMachineFireOccurance(pub u8);
 
 impl StateMachineFireOccurance {
     pub const AT_START: Self = Self(0);
@@ -30,7 +30,7 @@ pub trait StateMachineFireActionBehavior {
 
 impl StateMachineFireAction {
     pub fn occurs(&self) -> StateMachineFireOccurance {
-        StateMachineFireOccurance(self.base.occurs_value() as i32)
+        StateMachineFireOccurance(self.base.occurs_value())
     }
 
     pub fn import(&mut self, import_stack: &mut ImportStack) -> StatusCode {
