+++ crates/nuxie-runtime/src/mechanical_port/source/shapes/paint/stroke.rs
@@ -55,8 +55,8 @@ impl Stroke {
             return false;
         }
         let thickness = self.base.thickness();
-        let cap: nuxie_render_api::StrokeCap = StrokeCap::from(self.base.cap()).into();
-        let join: nuxie_render_api::StrokeJoin = StrokeJoin::from(self.base.join()).into();
+        let cap: nuxie_render_api::StrokeCap = StrokeCap::from(u32::from(self.base.cap())).into();
+        let join: nuxie_render_api::StrokeJoin = StrokeJoin::from(u32::from(self.base.join())).into();
         self.base.with_render_paint_mut(|paint| {
             paint.style(RenderPaintStyle::Stroke);
             paint.thickness(thickness);
@@ -69,8 +69,8 @@ impl Stroke {
     pub fn apply_to(&mut self, paint: &mut dyn RenderPaint, opacity: f32) {
         paint.style(RenderPaintStyle::Stroke);
         paint.thickness(self.base.thickness());
-        paint.cap(StrokeCap::from(self.base.cap()).into());
-        paint.join(StrokeJoin::from(self.base.join()).into());
+        paint.cap(StrokeCap::from(u32::from(self.base.cap())).into());
+        paint.join(StrokeJoin::from(u32::from(self.base.join())).into());
         paint.shader(None);
         let path_flags = self.path_flags();
         if let Some(mutator) = self.base.paint() {
@@ -100,12 +100,12 @@ impl Stroke {
 
     pub fn update(&mut self, value: ComponentDirt) {
         let kind = self.pick_path_kind();
-        let stroke = Some((self.base.thickness(), self.base.cap(), self.base.join()));
+        let stroke = Some((self.base.thickness(), u32::from(self.base.cap()), u32::from(self.base.join())));
         self.base.base.update_with_path_kind(value, kind, stroke);
         if has_dirt(value, ComponentDirt::PAINT) {
             let thickness = self.base.thickness();
-            let cap = StrokeCap::from(self.base.cap()).into();
-            let join = StrokeJoin::from(self.base.join()).into();
+            let cap = StrokeCap::from(u32::from(self.base.cap())).into();
+            let join = StrokeJoin::from(u32::from(self.base.join())).into();
             self.base.with_render_paint_mut(|paint| {
                 paint.thickness(thickness);
                 paint.cap(cap);
